+ Dodano menadżera materiałów. Towrzy on teksturę zawierającą materiały. + shadery umieją już wyświetlać kolory na podstawie danych z tekstury - generowanie dużej tekstury materiałów, tak żeby miliard planet nie odwoływało się do tej samej pamięci - planety powinny być źródłami świateł - FIXME: wczytywanie model_id jest zjebane
git-svn-id: svn+ssh://gamma/home/samba/koturj/svn/inz@110 e70bc5d4-e669-4a26-a90f-bfde2a246bde

+++ src/mem/misc/gfx_planet_factory.cpp
@@ -46,7 +46,7 @@ const GfxPlanet GfxPlanetFactory::getPlanet( int id ) const
 	return GfxPlanet( id , holder );
 }
 
-const BufferGl<float> &GfxPlanetFactory::getModels() const
+const BufferGl<int> &GfxPlanetFactory::getModels() const
 {
 	return holder->model;
 }
